requirements.txt
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<3.0.0
PyYAML>=6.0.0,<7.0.0
click>=8.1.0,<9.0.0
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<0.14.0
